pyproject.toml
[project]
name = "nakobot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pytelegrambotapi>=4.29.1",
    "python-dotenv>=1.2.1",
]
